requirements.txt
flask>=3.0.0,<4
google-ads>=28.0.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
